requirements.txt
Pillow>=8.0.0
tkcalendar>=1.6.0
